feat: move e2e test in cookie-consent
